well not it doenst crash any more but it laggs a bit
im super pist that this toom more that 15 mins.  i was pointing the "get
all references of an actor (which outputs an array) to a length function
and i couldnt figure out why it wouldnt work.  Apparently and array is
not an array.... who knew?

the program is now crashing

crashes after just a few colisions
balls
spawning is neemingly random places (as intended)